graph/requirements.txt
google-cloud-storage==1.7.0
requests==2.13.0
matplotlib==3.2.1
numpy
